fix index data types

+++ BasicRenderer/include/MeshUtilities.h
@@ -14,5 +14,5 @@ TangentBitangent calculateTangentsBitangentsIndexed(
     const std::vector<XMFLOAT3>& positions,
     const std::vector<XMFLOAT3>& normals,
     const std::vector<XMFLOAT2>& uvs,
-    const std::vector<uint16_t>& indices
+    const std::vector<uint32_t>& indices
 );
